pyproject.toml
[tool.poetry]
name = "argus"
version = "0.1.0"
description = "argus using crewAI"
authors = [ "Renzo Barrios <renzo@devpulse.xyz>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = "^0.51.0" }
pillow = "^10.4.0"
google-generativeai = "^0.7.2"
langchain-google-genai = "^1.0.8"
langchain = "^0.2.14"
python-dotenv = "^1.0.1"
openai = "^1.40.8"
requests = "^2.32.3"
ipython = "^8.26.0"
fastapi = {extras = ["standard"], version = "^0.112.1"}
fastapi-versionizer = "^4.0.1"

[tool.poetry.scripts]
argus = "argus.main:run"
run_crew = "argus.main:run"
train = "argus.main:train"
replay = "argus.main:replay"
test = "argus.main:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
